logging:  harmonise disabling of tracing
1. --tracing-off and --log-config are now mutually exclusive
2. all users now respect --tracing-off
3. some code shuffling for better locality
